pyproject.toml
[project]
name = "qwen3-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "einops>=0.8.1",
    "jaxtyping>=0.3.3",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pytest>=8.4.2",
    "pyyaml>=6.0.3",
    "tiktoken>=0.12.0",
]
